Poprawki int na long id Dodanie do api Category

+++ renter-api/Datebase/Interfaces/IMovieRatingRepositoryService.cs
@@ -7,7 +7,7 @@ namespace Database.Interfaces
 {
     public interface IMovieRatingRepositoryService : IRepositoryService<MovieRating>
     {
-        float GetRate(int movieId);
-        void RateMovie(int movieId, string userId, int rate);
+        float GetRate(long movieId);
+        void RateMovie(long movieId, string userId, int rate);
     }
 }
